Spectral canonization

+++ include/kitty/static_truth_table.hpp
@@ -58,7 +58,7 @@ struct static_truth_table<NumVars, true>
   /*! \cond PRIVATE */
   enum
   {
-    NumBits = 1 << NumVars
+    NumBits = uint64_t( 1 ) << NumVars
   };
   /*! \endcond */
 
@@ -139,7 +139,7 @@ struct static_truth_table<NumVars, false>
 
   enum
   {
-    NumBits = 1 << NumVars
+    NumBits = uint64_t( 1 ) << NumVars
   };
   /*! \endcond */
